Fixed decoding on resume.
After successfully reading a barcode, the app may be suspended to the
background. When the app is then called back, it will only show the
preview, but it will not try to decode. However, if the app did not find
a code, it will restart the decoding. This behaviour can be manipulated
to always forcefully restart the decoding on resume.